:sparkles: BaseViewController로 이메일 발송 기능 이관 및 신고 양식 개발
